dev: Break up script.py into smaller more manageable files
